Moved account data functions into mixin for re-use by the billing page
